V4.x

+++ app/src/Controller/Admin/Pages.php
@@ -2,16 +2,12 @@
 
 namespace Dappur\Controller\Admin;
 
-use Carbon\Carbon;
 use Dappur\Controller\Controller as Controller;
-use Dappur\Dappurware\FileResponse;
-use Dappur\Dappurware\Utils;
 use Psr\Http\Message\ServerRequestInterface as Request;
 use Psr\Http\Message\ResponseInterface as Response;
-use Respect\Validation\Validator as V;
 
 /**
- * @SuppressWarnings(PHPMD.StaticAccess)
+ * @SuppressWarnings(PHPMD.ExcessiveClassComplexity)
  */
 class Pages extends Controller
 {
@@ -20,8 +16,8 @@ class Pages extends Controller
         if ($check = $this->sentinel->hasPerm('pages.view', 'dashboard')) {
             return $check;
         }
-  
-        $totalData = \Dappur\Model\Routes::count();
+        $routes = new \Dappur\Model\Routes;
+        $totalData = $routes->count();
             
         $totalFiltered = $totalData;
 
@@ -84,16 +80,17 @@ class Pages extends Controller
         if ($check = $this->sentinel->hasPerm('pages.view', 'dashboard')) {
             return $check;
         }
-
+        $roles = new \Dappur\Model\Roles;
         return $this->view->render(
             $response,
             'pages.twig',
-            ["roles" => \Dappur\Model\Roles::get()]
+            ["roles" => $roles->get()]
         );
     }
 
     /**
-     * @SuppressWarnings(PHPMD.UnusedFormalParameter)
+     * @SuppressWarnings(PHPMD.CyclomaticComplexity)
+     * @SuppressWarnings(PHPMD.NPathComplexity)
      */
     public function add(Request $request, Response $response)
     {
@@ -156,16 +153,17 @@ class Pages extends Controller
             $this->flash('success', $insPage->name . " has been successfully created.");
             return $this->redirect($response, 'admin-pages');
         }
-
+        $roles = new \Dappur\Model\Roles;
         return $this->view->render(
             $response,
             'pages-add.twig',
-            ["roles" => \Dappur\Model\Roles::get()]
+            ["roles" => $roles->get()]
         );
     }
 
     /**
-     * @SuppressWarnings(PHPMD.UnusedFormalParameter)
+     * @SuppressWarnings(PHPMD.CyclomaticComplexity)
+     * @SuppressWarnings(PHPMD.NPathComplexity)
      */
     public function edit(Request $request, Response $response, $routeId)
     {
@@ -192,8 +190,9 @@ class Pages extends Controller
                 $this->validator->addError('name', 'Name already exists');
             }
 
-            if (in_array("/" . $request->getParam('pattern'), $routePatterns) && $routeCheck->pattern != $request->getParam('pattern')) {
-                $this->validator->addError('pattern', 'Pattern already exists.');
+            if (in_array("/" . $request->getParam('pattern'), $routePatterns)
+                && $routeCheck->pattern != $request->getParam('pattern')) {
+                    $this->validator->addError('pattern', 'Pattern already exists.');
             }
             
             $routeCheck->name = $request->getParam('name');
@@ -241,11 +240,11 @@ class Pages extends Controller
             $this->flash('success', $routeCheck->name . " has been successfully updated.");
             return $this->redirect($response, 'admin-pages');
         }
-
+        $roles = new \Dappur\Model\Roles;
         return $this->view->render(
             $response,
             'pages-edit.twig',
-            ["roles" => \Dappur\Model\Roles::get(), "route" => $routeCheck]
+            ["roles" => $roles->get(), "route" => $routeCheck]
         );
     }
 
@@ -290,7 +289,8 @@ class Pages extends Controller
         fwrite($tempFile, json_encode($final, JSON_PRETTY_PRINT));
         $metaDatas = stream_get_meta_data($tempFile);
         $filePath = $metaDatas['uri'];
-        return FileResponse::getResponse(
+        $fileResponse = new \Dappur\Dappurware\FileResponse;
+        return $fileResponse->getResponse(
             $response,
             $filePath,
             $this->settings['framework'] .
@@ -422,7 +422,7 @@ class Pages extends Controller
         // Process Config Items
         foreach ($roles as $role) {
             if ($overwrite) {
-                $deleteRoles = \Dappur\Model\RoleRoutes::where('route_id', $route->id)
+                \Dappur\Model\RoleRoutes::where('route_id', $route->id)
                     ->get()
                     ->delete();
             }
@@ -448,8 +448,13 @@ class Pages extends Controller
         if ($check = $this->sentinel->hasPerm('pages.delete', 'dashboard')) {
             return $check;
         }
+        $routes = new \Dappur\Model\Routes;
+        $page = $routes->find($request->getParam('page_id'));
 
-        $page = \Dappur\Model\Routes::find($request->getParam('page_id'));
+        if ($page && $page->id == 1) {
+            $this->flash('danger', 'You cannot delete the home page.');
+            return $this->redirect($response, 'admin-pages');
+        }
 
         if ($page->delete()) {
             $this->flash('success', 'Page has been deleted successfully.');
